+++ src/hooks/ready/registerGameSettings.js
@@ -516,6 +516,15 @@ export default function () {
     default: false,
   });
 
+  game.settings.register("ddb-importer", "character-update-policy-active-effect-character-copy", {
+    name: "ddb-importer.character-update-policy-active-effect-character-copy.name",
+    hint: "ddb-importer.character-update-policy-active-effect-character-copy.hint",
+    scope: "player",
+    config: false,
+    type: Boolean,
+    default: false,
+  });
+
   /** Munching update settings, stored per user and non-configurable in the settings screen */
   game.settings.register("ddb-importer", "munching-policy-update-existing", {
     name: "ddb-importer.munching-policy-update-existing.name",